+++ packages/client/src/devices/MicrophoneManager.ts
@@ -51,7 +51,7 @@ export class MicrophoneManager extends AudioDeviceManager<MicrophoneManagerState
   ) {
     super(
       call,
-      new MicrophoneManagerState(disableMode),
+      new MicrophoneManagerState(disableMode, call.tracer),
       TrackType.AUDIO,
       devicePersistence,
     );
@@ -169,7 +169,6 @@ export class MicrophoneManager extends AudioDeviceManager<MicrophoneManagerState
                 deviceId,
                 label,
               };
-              console.log(event);
               this.call.tracer.trace('mic.capture_report', event);
               this.call.streamClient.dispatchEvent(event);
             },