+++ sample-apps/react/react-dogfood/components/ToggleNoiseCancellationButton.tsx
@@ -1,19 +1,31 @@
 import {
   CompositeButton,
   Icon,
+  LoadingIndicator,
   WithTooltip,
   useNoiseCancellation,
 } from '@stream-io/video-react-sdk';
 
 export const ToggleNoiseCancellationButton = () => {
-  const { isSupported, isEnabled, setEnabled } = useNoiseCancellation();
+  const { isSupported, isEnabled, isReady, setEnabled } =
+    useNoiseCancellation();
+
   if (!isSupported) return null;
+
   return (
     <WithTooltip
-      title={`Noise cancellation is ${isEnabled ? 'active' : 'inactive'}`}
+      title={`Noise cancellation is ${!isReady ? 'loading' : isEnabled ? 'active' : 'inactive'}`}
     >
-      <CompositeButton onClick={() => setEnabled((v) => !v)} variant="primary">
-        <Icon icon={isEnabled ? 'anc' : 'anc-off'} />
+      <CompositeButton
+        disabled={!isReady}
+        onClick={() => setEnabled((v) => !v)}
+        variant="primary"
+      >
+        {!isReady ? (
+          <LoadingIndicator />
+        ) : (
+          <Icon icon={isEnabled ? 'anc' : 'anc-off'} />
+        )}
       </CompositeButton>
     </WithTooltip>
   );